[ValueTrackingUtils] Split out from ValueTracking (NFC)
ValueTracking has grown in size over the years, and is currently huge.
In order to split it, we have picked all routines that do not directly
or indirectly depend on computeKnownBits and moved them into a new file:
this was relatively straightfoward, as all routines dependent on
computeKnownBits must have a Depth parameter.

We have chosen not to expose ValueTrackingUtils.h in order to avoid
blowing up the public include directory.
